+++ e2e/cypress/integration/channel/user_to_channel_admin_member_updates_manage_channel_members_modal_spec.js
@@ -71,7 +71,7 @@ describe('Change Roles', () => {
         });
     });
 
-    it('MM-10858 - Going from a Channel Member to Channel Admin update view member modal without refresh', () => {
+    it('MM-T4174 User role to channel admin/member updates channel member modal immediately without refresh', () => {
         // # Go to member modal
         cy.get('#member_popover').click();
         cy.findByTestId('membersModal').click();
